+++ certbot/cli.py
@@ -101,6 +101,7 @@ manage certificates:
 
 manage your account with Let's Encrypt:
     register        Create a Let's Encrypt ACME account
+    unregister      Deactivate a Let's Encrypt ACME account
     update_account  Update a Let's Encrypt ACME account
   --agree-tos       Agree to the ACME server's Subscriber Agreement
    -m EMAIL         Email address for important account notifications
@@ -117,7 +118,7 @@ More detailed help:
    all, automation, commands, paths, security, testing, or any of the
    subcommands or plugins (certonly, renew, install, register, nginx,
    apache, standalone, webroot, etc.)
-
+  -h all                print a detailed help page including all topics 
   --version             print the version number
 """
 
@@ -311,9 +312,8 @@ def config_help(name, hidden=False):
     # pylint: disable=no-member
     if hidden:
         return argparse.SUPPRESS
-    else:
-        field = interfaces.IConfig.__getitem__(name) # type: zope.interface.interface.Attribute
-        return field.__doc__
+    field = interfaces.IConfig.__getitem__(name)  # type: zope.interface.interface.Attribute  # pylint: disable=no-value-for-parameter
+    return field.__doc__
 
 
 class HelpfulArgumentGroup(object):
@@ -539,7 +539,7 @@ class HelpfulArgumentParser(object):
     # Help that are synonyms for --help subcommands
     COMMANDS_TOPICS = ["command", "commands", "subcommand", "subcommands", "verbs"]
     def _list_subcommands(self):
-        longest = max(len(v) for v in VERB_HELP_MAP.keys())
+        longest = max(len(v) for v in VERB_HELP_MAP)
 
         text = "The full list of available SUBCOMMANDS is:\n\n"
         for verb, props in sorted(VERB_HELP):
@@ -567,7 +567,7 @@ class HelpfulArgumentParser(object):
             apache_doc = "(the certbot apache plugin is not installed)"
 
         usage = SHORT_USAGE
-        if help_arg == True:
+        if help_arg is True:
             self.notify(usage + COMMAND_OVERVIEW % (apache_doc, nginx_doc) + HELP_AND_VERSION_USAGE)
             sys.exit(0)
         elif help_arg in self.COMMANDS_TOPICS:
@@ -873,8 +873,7 @@ class HelpfulArgumentParser(object):
                          for t in self.help_topics])
         elif not chosen_topic:
             return dict([(t, False) for t in self.help_topics])
-        else:
-            return dict([(t, t == chosen_topic) for t in self.help_topics])
+        return dict([(t, t == chosen_topic) for t in self.help_topics])
 
 def _add_all_groups(helpful):
     helpful.add_group("automation", description="Flags for automating execution & other tweaks")
@@ -1118,14 +1117,6 @@ def prepare_and_parse_args(plugins, args, detect_defaults=False):  # pylint: dis
         help=config_help("no_verify_ssl"),
         default=flag_default("no_verify_ssl"))
     helpful.add(
-        ["testing", "standalone", "apache", "nginx"], "--tls-sni-01-port", type=int,
-        default=flag_default("tls_sni_01_port"),
-        help=config_help("tls_sni_01_port"))
-    helpful.add(
-        ["testing", "standalone"], "--tls-sni-01-address",
-        default=flag_default("tls_sni_01_address"),
-        help=config_help("tls_sni_01_address"))
-    helpful.add(
         ["testing", "standalone", "manual"], "--http-01-port", type=int,
         dest="http01_port",
         default=flag_default("http01_port"), help=config_help("http01_port"))
@@ -1133,6 +1124,10 @@ def prepare_and_parse_args(plugins, args, detect_defaults=False):  # pylint: dis
         ["testing", "standalone"], "--http-01-address",
         dest="http01_address",
         default=flag_default("http01_address"), help=config_help("http01_address"))
+    helpful.add(
+        ["testing", "nginx"], "--https-port", type=int,
+        default=flag_default("https_port"),
+        help=config_help("https_port"))
     helpful.add(
         "testing", "--break-my-certs", action="store_true",
         default=flag_default("break_my_certs"),
@@ -1193,7 +1188,7 @@ def prepare_and_parse_args(plugins, args, detect_defaults=False):  # pylint: dis
         action=_PrefChallAction, default=flag_default("pref_challs"),
         help='A sorted, comma delimited list of the preferred challenge to '
              'use during authorization with the most preferred challenge '
-             'listed first (Eg, "dns" or "tls-sni-01,http,dns"). '
+             'listed first (Eg, "dns" or "http,dns"). '
              'Not all plugins support all challenges. See '
              'https://certbot.eff.org/docs/using.html#plugins for details. '
              'ACME Challenges are versioned, but if you pick "http" rather '
@@ -1263,6 +1258,17 @@ def prepare_and_parse_args(plugins, args, detect_defaults=False):  # pylint: dis
 
     helpful.add_deprecated_argument("--agree-dev-preview", 0)
     helpful.add_deprecated_argument("--dialog", 0)
+
+    # Deprecation of tls-sni-01 related cli flags
+    # TODO: remove theses flags completely in few releases
+    class _DeprecatedTLSSNIAction(util._ShowWarning):  # pylint: disable=protected-access
+        def __call__(self, parser, namespace, values, option_string=None):
+            super(_DeprecatedTLSSNIAction, self).__call__(parser, namespace, values, option_string)
+            namespace.https_port = values
+    helpful.add(
+        ["testing", "standalone", "apache", "nginx"], "--tls-sni-01-port",
+        type=int, action=_DeprecatedTLSSNIAction, help=argparse.SUPPRESS)
+    helpful.add_deprecated_argument("--tls-sni-01-address", 1)
 
     # Populate the command line parameters for new style enhancements
     enhancements.populate_cli(helpful.add)
@@ -1556,6 +1562,15 @@ def parse_preferred_challenges(pref_challs):
     aliases = {"dns": "dns-01", "http": "http-01", "tls-sni": "tls-sni-01"}
     challs = [c.strip() for c in pref_challs]
     challs = [aliases.get(c, c) for c in challs]
+
+    # Ignore tls-sni-01 from the list, and generates a deprecation warning
+    # TODO: remove this option completely in few releases
+    if "tls-sni-01" in challs:
+        logger.warning('TLS-SNI-01 support is deprecated. This value is being dropped from the '
+                       'setting of --preferred-challenges and future versions of Certbot will '
+                       'error if it is included.')
+        challs = [chall for chall in challs if chall != "tls-sni-01"]
+
     unrecognized = ", ".join(name for name in challs
                              if name not in challenges.Challenge.TYPES)
     if unrecognized:
@@ -1563,10 +1578,12 @@ def parse_preferred_challenges(pref_challs):
             "Unrecognized challenges: {0}".format(unrecognized))
     return challs
 
+
 def _user_agent_comment_type(value):
     if "(" in value or ")" in value:
         raise argparse.ArgumentTypeError("may not contain parentheses")
     return value
+
 
 class _DeployHookAction(argparse.Action):
     """Action class for parsing deploy hooks."""