+++ certbot/plugins/selection_test.py
@@ -6,10 +6,10 @@ import unittest
 import mock
 import zope.component
 
+from acme.magic_typing import List  # pylint: disable=unused-import, no-name-in-module
+
 from certbot import errors
 from certbot import interfaces
-
-from acme.magic_typing import List  # pylint: disable=unused-import, no-name-in-module
 from certbot.display import util as display_util
 from certbot.plugins.disco import PluginsRegistry
 from certbot.tests import util as test_util