+++ certbot/plugins/standalone_test.py
@@ -1,9 +1,8 @@
 """Tests for certbot.plugins.standalone."""
-import argparse
 import socket
-import unittest
 # https://github.com/python/typeshed/blob/master/stdlib/2and3/socket.pyi
 from socket import errno as socket_errors  # type: ignore
+import unittest
 
 import josepy as jose
 import mock
@@ -44,9 +43,6 @@ class ServerManagerTest(unittest.TestCase):
         self.mgr.stop(port=port)
         self.assertEqual(self.mgr.running(), {})
 
-    def test_run_stop_tls_sni_01(self):
-        self._test_run_stop(challenges.TLSSNI01)
-
     def test_run_stop_http_01(self):
         self._test_run_stop(challenges.HTTP01)
 
@@ -76,46 +72,6 @@ class ServerManagerTest(unittest.TestCase):
         maybe_another_server.close()
 
 
-class SupportedChallengesActionTest(unittest.TestCase):
-    """Tests for plugins.standalone.SupportedChallengesAction."""
-
-    def _call(self, value):
-        with mock.patch("certbot.plugins.standalone.logger") as mock_logger:
-            # stderr is mocked to prevent potential argparse error
-            # output from cluttering test output
-            with mock.patch("sys.stderr"):
-                config = self.parser.parse_args([self.flag, value])
-
-        self.assertTrue(mock_logger.warning.called)
-        return getattr(config, self.dest)
-
-    def setUp(self):
-        self.flag = "--standalone-supported-challenges"
-        self.dest = self.flag[2:].replace("-", "_")
-        self.parser = argparse.ArgumentParser()
-
-        from certbot.plugins.standalone import SupportedChallengesAction
-        self.parser.add_argument(self.flag, action=SupportedChallengesAction)
-
-    def test_correct(self):
-        self.assertEqual("tls-sni-01", self._call("tls-sni-01"))
-        self.assertEqual("http-01", self._call("http-01"))
-        self.assertEqual("tls-sni-01,http-01", self._call("tls-sni-01,http-01"))
-        self.assertEqual("http-01,tls-sni-01", self._call("http-01,tls-sni-01"))
-
-    def test_unrecognized(self):
-        assert "foo" not in challenges.Challenge.TYPES
-        self.assertRaises(SystemExit, self._call, "foo")
-
-    def test_not_subset(self):
-        self.assertRaises(SystemExit, self._call, "dns")
-
-    def test_dvsni(self):
-        self.assertEqual("tls-sni-01", self._call("dvsni"))
-        self.assertEqual("http-01,tls-sni-01", self._call("http-01,dvsni"))
-        self.assertEqual("tls-sni-01,http-01", self._call("dvsni,http-01"))
-
-
 def get_open_port():
     """Gets an open port number from the OS."""
     open_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM, 0)
@@ -131,32 +87,16 @@ class AuthenticatorTest(unittest.TestCase):
     def setUp(self):
         from certbot.plugins.standalone import Authenticator
 
-        self.config = mock.MagicMock(
-            tls_sni_01_port=get_open_port(), http01_port=get_open_port(),
-            standalone_supported_challenges="tls-sni-01,http-01")
+        self.config = mock.MagicMock(http01_port=get_open_port())
         self.auth = Authenticator(self.config, name="standalone")
         self.auth.servers = mock.MagicMock()
-
-    def test_supported_challenges(self):
-        self.assertEqual(self.auth.supported_challenges,
-                         [challenges.TLSSNI01, challenges.HTTP01])
-
-    def test_supported_challenges_configured(self):
-        self.config.standalone_supported_challenges = "tls-sni-01"
-        self.assertEqual(self.auth.supported_challenges,
-                         [challenges.TLSSNI01])
 
     def test_more_info(self):
         self.assertTrue(isinstance(self.auth.more_info(), six.string_types))
 
     def test_get_chall_pref(self):
         self.assertEqual(self.auth.get_chall_pref(domain=None),
-                         [challenges.TLSSNI01, challenges.HTTP01])
-
-    def test_get_chall_pref_configured(self):
-        self.config.standalone_supported_challenges = "tls-sni-01"
-        self.assertEqual(self.auth.get_chall_pref(domain=None),
-                         [challenges.TLSSNI01])
+                         [challenges.HTTP01])
 
     def test_perform(self):
         achalls = self._get_achalls()
@@ -212,10 +152,8 @@ class AuthenticatorTest(unittest.TestCase):
         key = jose.JWK.load(test_util.load_vector('rsa512_key.pem'))
         http_01 = achallenges.KeyAuthorizationAnnotatedChallenge(
             challb=acme_util.HTTP01_P, domain=domain, account_key=key)
-        tls_sni_01 = achallenges.KeyAuthorizationAnnotatedChallenge(
-            challb=acme_util.TLSSNI01_P, domain=domain, account_key=key)
 
-        return [http_01, tls_sni_01]
+        return [http_01]
 
     def test_cleanup(self):
         self.auth.servers.running.return_value = {
@@ -242,6 +180,7 @@ class AuthenticatorTest(unittest.TestCase):
         self.assertEqual(self.auth.served, {
             "server1": set(), "server2": set([])})
         self.auth.servers.stop.assert_called_with(2)
+
 
 if __name__ == "__main__":
     unittest.main()  # pragma: no cover