+++ tools/venv.py
@@ -1,41 +1,32 @@
 #!/usr/bin/env python
 # Developer virtualenv setup for Certbot client
 import os
+import sys
 
 import _venv_common
 
-REQUIREMENTS = [
-    '-e acme[dev]',
-    '-e .[dev,docs]',
-    '-e certbot-apache',
-    '-e certbot-dns-cloudflare',
-    '-e certbot-dns-cloudxns',
-    '-e certbot-dns-digitalocean',
-    '-e certbot-dns-dnsimple',
-    '-e certbot-dns-dnsmadeeasy',
-    '-e certbot-dns-gehirn',
-    '-e certbot-dns-google',
-    '-e certbot-dns-linode',
-    '-e certbot-dns-luadns',
-    '-e certbot-dns-nsone',
-    '-e certbot-dns-ovh',
-    '-e certbot-dns-rfc2136',
-    '-e certbot-dns-route53',
-    '-e certbot-dns-sakuracloud',
-    '-e certbot-nginx',
-    '-e certbot-postfix',
-    '-e letshelp-certbot',
-    '-e certbot-compatibility-test',
-]
+def create_venv(venv_path):
+    """Create a Python 2 virtual environment at venv_path.
+
+    :param str venv_path: path where the venv should be created
+
+    """
+    python2 = _venv_common.find_python_executable(2)
+    command = [sys.executable, '-m', 'virtualenv', '--python', python2, venv_path]
+
+    environ = os.environ.copy()
+    environ['VIRTUALENV_NO_DOWNLOAD'] = '1'
+    _venv_common.subprocess_with_print(command, environ)
 
 
-def main():
+def main(pip_args=None):
     if os.name == 'nt':
         raise ValueError('Certbot for Windows is not supported on Python 2.x.')
 
-    venv_args = '--python "{0}"'.format(_venv_common.find_python_executable(2))
-    _venv_common.main('venv', venv_args, REQUIREMENTS)
+    venv_path = _venv_common.prepare_venv_path('venv')
+    create_venv(venv_path)
+    _venv_common.install_packages(venv_path, pip_args)
 
 
 if __name__ == '__main__':
-    main()
+    main(sys.argv[1:])